feat: added plug for empty list and refactored utils.js
