+++ src/utils/event.js
@@ -42,11 +42,4 @@ const humanizeDateForEvent = (date) => date ? dayjs(date).utc().format(DATE_FORM
 const humanizeTimeFrom = (date) => date ? dayjs(date).utc().format(DATE_FORMAT_FOR_EVENT_TIME) : '';
 const humanizeTimeTo = (date) => date ? dayjs(date).utc().format(DATE_FORMAT_FOR_EVENT_TIME) : '';
 
-
-// другие утилитарные функции
-const getRandomArrayElement = (items) => items[Math.floor(Math.random() * items.length)];
-
-const getRandomInteger = (min, max) => Math.floor(Math.random() * (Math.floor(max) - Math.ceil(min) + 1)) + min;
-
-// console.log(humanizeDate('2019-07-10T22:55:56.845Z'));
-export {getRandomArrayElement, getRandomInteger, humanizeDateForEdit, humanizeDateForEvent, humanizeTimeFrom, humanizeTimeTo, getTimeGap};
+export {humanizeDateForEdit, humanizeDateForEvent, humanizeTimeFrom, humanizeTimeTo, getTimeGap};
