+++ src/view/trip-event-edit-view.js
@@ -1,5 +1,5 @@
 import AbstractView from '../framework/view/abstract-view.js';
-import { humanizeDateForEdit } from '../util.js';
+import { humanizeDateForEdit } from '../utils/event.js';
 import { WAYPOINT_TYPES, DESTINATIONS_NAMES } from '../const.js';
 
 const BLANK_EVENT = {
